+++ src/lib/quests/index.ts
@@ -1,2 +1,3 @@
 export * from "./types";
 export * from "./transitions";
+export * from "./matchSkill";